requirements.txt
pydantic
langchain
langgraph
langchain-community
python-dotenv
langchain-groq
arxiv
wikipedia
fastapi
uvicorn
